requirements.txt
akshare
pandas
numpy
matplotlib
requests
yfinance
openpyxl
pillow
pykrx
pandas_market_calendars
